Docker support, switch to pgsql, refactor, update neo4j

+++ src/main/java/net/simon987/musicgraph/entities/Relation.java
@@ -2,7 +2,7 @@ package net.simon987.musicgraph.entities;
 
 
 public class Relation {
-    public float weight;
+    public double weight;
     public long source;
     public long target;
 }
